Bugfix in response import and small bugfixes
Hotfix of linelist ID (ArII -> HeI), shared axes deprecation fix in
identify_gui.py, and bug in loading filelists of length 1.
