+++ pynot/VERSION
@@ -1,1 +1,1 @@
-2.0.1
+2.1
