+++ pynot/calib/HeAr_linelist.dat
@@ -10,7 +10,7 @@
 4471.479   HeI 
 4510.733   ArI 
 4713.146   HeI 
-4921.931   ArII
+4921.931   HeI
 5015.680   HeI 
 5875.621   HeI 
 6114.923   ArII
