+++ pynot/response.py
@@ -26,7 +26,7 @@ from pynot import instrument
 from pynot.data import organizer
 from pynot.extraction import auto_extract
 from pynot import extract_gui
-from pynot.functions import get_version_number, my_formatter, mad
+from pynot.functions import get_version_number, my_formatter, mad, lookup_std_star
 from pynot import response_gui
 from pynot.logging import Report
 from pynot.scired import raw_correction
@@ -44,33 +44,6 @@ _standard_star_files = glob.glob(path + '/calib/std/*.dat')
 _standard_star_files = [basename(fname) for fname in _standard_star_files]
 # List of star names in lowercase:
 standard_stars = [fname.strip('.dat') for fname in _standard_star_files]
-
-# Look-up table from target-names -> star names
-# (mostly used for ALFOSC where TCSTGT is different)
-std_fname = os.path.join(path, 'calib/std/tcs_namelist.txt')
-calib_names = np.loadtxt(std_fname, dtype=str, delimiter=':')
-tcs_standard_stars = {row[1].strip(): row[0].strip() for row in calib_names}
-
-
-def lookup_std_star(hdr):
-    """
-    Check if the given header contains an object or target name
-    which matches one of the defined standard calibration stars.
-
-    Returns `None` if no match is found.
-    """
-    object_name = instrument.get_object(hdr).replace(' ', '')
-    target_name = instrument.get_target_name(hdr).replace(' ', '')
-    if object_name.lower() in standard_stars:
-        return object_name.lower()
-    elif object_name.upper() in tcs_standard_stars:
-        return tcs_standard_stars[object_name.upper()]
-    elif target_name.lower() in standard_stars:
-        return target_name.lower()
-    elif target_name.upper() in tcs_standard_stars:
-        return tcs_standard_stars[target_name.upper()]
-    else:
-        return None
 
 
 def load_spectrum1d(fname):
